+++ src/doomstat.c
@@ -202,6 +202,7 @@ int show_save_messages; // CFG-Only
 int show_ssg; // CFG-Only
 int hud_stats_format;
 int hud_stats_format_map;
+int hud_stats_icons;
 int alt_arms;
 int hud_time_teleport;
 int hud_time_keypickup;