+++ src/doomstat.h
@@ -564,6 +564,7 @@ enum {
 };
 extern int hud_stats_format;
 extern int hud_stats_format_map;
+extern int hud_stats_icons;
 
 extern int alt_arms;